fix: 자동차 이동 횟수 입력 유효성 검사 추가

+++ src/main/kotlin/racingcar/controller/RacingCarController.kt
@@ -14,7 +14,7 @@ class RacingCarController(
         val names = racingCarService.parseCarNameInput(carNamesInput)
         val cars = racingCarService.getRacingPlayers(names)
         racingCarView.printCarRaceResultTitle()
-        repeat(carCountInput.toInt()) {
+        repeat(carCountInput) {
             racingCarService.playRace(cars)
             racingCarView.printCarRaceCurrentResult(cars)
         }
